+++ src/main/java/org/mongeez/ChangeSetExecutor.java
@@ -1,5 +1,6 @@
 /*
  * Copyright 2011 SecondMarket Labs, LLC.
+ * Copyright 2023 Hitesh Tarani
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,10 +17,16 @@ import org.mongeez.commands.ChangeSet;
 import org.mongeez.commands.Script;
 import org.mongeez.dao.MongeezDao;
 
-import com.mongodb.Mongo;
+import com.mongodb.MongoClient;
+import org.mongeez.dao.MongoShellScriptExecutor;
+import org.mongeez.dao.ShellScriptExecutor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.List;
 
 
@@ -29,12 +36,22 @@ public class ChangeSetExecutor {
     private MongeezDao dao = null;
     private String context = null;
 
-    public ChangeSetExecutor(Mongo mongo, String dbName, String context) {
-        this(mongo, dbName, context, null);
+    /**
+     * This executes scripts with a {@link org.mongeez.dao.MongoShellScriptExecutor} which needs
+     * <a href="https://www.mongodb.com/docs/v4.4/mongo/">mongo shell client</a> executable available on the system path.
+     * To use the alternative mongo shells, use {@link ChangeSetExecutor#ChangeSetExecutor(com.mongodb.MongoClient, java.lang.String, java.lang.String, org.mongeez.dao.ShellScriptExecutor)}
+     */
+    public ChangeSetExecutor(MongoClient mongo, String mongoClientUri, String context, String dbName) {
+        dao = new MongeezDao(mongo, dbName, new MongoShellScriptExecutor(mongoClientUri));
+        this.context = context;
     }
 
-    public ChangeSetExecutor(Mongo mongo, String dbName, String context, MongoAuth auth) {
-        dao = new MongeezDao(mongo, dbName, auth);
+    /**
+     * This executes scripts with the provided shell client executor
+     * @see ShellScriptExecutor
+     */
+    public ChangeSetExecutor(MongoClient mongo, String context, String dbName, ShellScriptExecutor executor) {
+        dao = new MongeezDao(mongo, dbName, executor);
         this.context = context;
     }
 
@@ -55,17 +72,59 @@ public class ChangeSetExecutor {
     }
 
     private void execute(ChangeSet changeSet) {
+        File changesetTempFile = null;
         try {
             for (Script command : changeSet.getCommands()) {
-                command.run(dao);
+              changesetTempFile = File.createTempFile(getTempFilePath(changeSet), ".js");
+              writeChangeSetBodyToFile(command.getBody(), changesetTempFile.getPath());
+              dao.runScript(changesetTempFile.getAbsolutePath());
+              deleteFile(changesetTempFile);
             }
         } catch (RuntimeException e) {
+            deleteFile(changesetTempFile);
             if (changeSet.isFailOnError()) {
                 throw e;
             } else {
                 logger.warn("ChangeSet " + changeSet.getChangeId() + " has failed, but failOnError is set to false", e.getMessage());
             }
+        } catch (IOException e) {
+            deleteFile(changesetTempFile);
+            if (changeSet.isFailOnError()) {
+                throw new RuntimeException(e);
+            } else {
+                logger.warn("ChangeSet " + changeSet.getChangeId() + " has failed, but failOnError is set to false", e.getMessage());
+            }
         }
         dao.logChangeSet(changeSet);
+    }
+
+    private String getTempFilePath(ChangeSet changeSet) {
+      String resourcePath = changeSet.getResourcePath();
+      String changeFilePrefix = resourcePath.substring(0, resourcePath.lastIndexOf("."));
+      return changeFilePrefix + "." + changeSet.getChangeId();
+    }
+
+    private void writeChangeSetBodyToFile(String body, String filePath) {
+        PrintWriter out = null;
+  
+        try {
+          FileWriter fw = new FileWriter(filePath);
+          out = new PrintWriter(fw);
+          out.write(body);
+        } catch (IOException e) {
+            throw new RuntimeException("Unable to write to temporary changeset file" + filePath);
+        } finally {
+          if (out != null) {
+            out.close();
+          }
+        }
+      }
+
+    private void deleteFile(File file) {
+        try {
+            file.delete();
+        } catch (Exception e) {
+            logger.error("Unable to delete temporary file" + file.getPath());
+        }
     }
 }