+++ src/main/java/org/mongeez/dao/ShellScriptExecutor.java
@@ -1,0 +1,5 @@
+package org.mongeez.dao;
+
+public interface ShellScriptExecutor {
+  void runScript(String filePath);
+}